resume-app/requirements.txt
PyYAML
weasyprint
streamlit
pyyaml
python-dotenv
groq
weasyprint
